Fix sample project
This commit fixes some of the issues found in this sample project, more
specifically:

1. No dependency between demo and activityops so there is zero chance
that the controller can be found by the Spring Boot app
2. The controller was defined in the root package whereas Spring Boot
looks for `com.example` in this particular sample (the package in which
the `@SpringBootApplication` annotated class is located).
3. The root pom should not define the dependencies of the whole project
but should rather provide the necessary `dependencyManagement` for the
children projects.
4. Do not commit target and IDE-specific files and configure a
`.gitignore` for the project
