+++ multimodule/.gitignore
@@ -1,0 +1,3 @@
+target
+.idea
+*.iml
