+++ multimodule/activityops/src/main/java/com/example/test/User.java
@@ -1,0 +1,34 @@
+package com.example.test;
+
+public class User {
+
+    private Long id;
+
+    private String firstName;
+
+    private String lastName;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+}
