+++ multimodule/demo/pom.xml
@@ -5,10 +5,27 @@
     <parent>
         <artifactId>multimodule</artifactId>
         <groupId>org.multimodule</groupId>
-        <version>1.0</version>
+        <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>demo</artifactId>
 
+	<dependencies>
+		<dependency>
+			<groupId>org.multimodule</groupId>
+			<artifactId>activityops</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
+
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+	</dependencies>
 
 </project>
